requirement.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
Django==4.0
django-pdfkit==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
future==0.18.2
html5lib==1.1
pdfkit==0.6.0
Pillow==9.0.0
PyJWT==2.3.0
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.5
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
webencodings==0.5.1
